Changing Exchange url appearance and behaviour
Exchange url is pointing to SLM/BTC pair now, but displays only the Exchange's main url since the the SLM/BTC pair url is too long and doesn't have a nice appearance if put under the name of the Exchange. The url is also opening in the new tab now.